+++ litebox_syscall_rewriter/src/aarch64.rs
@@ -110,14 +110,20 @@
 //! traps during runtime rewriting.
 
 use alloc::format;
+use alloc::string::ToString as _;
 use alloc::vec::Vec;
 use core::ops::Range;
+use object::read::{Object as _, ObjectSection as _, ObjectSymbol as _};
 use yaxpeax_arch::{Decoder, U8Reader};
 use yaxpeax_arm::armv8::a64::{
     InstDecoder, Instruction as DecodedInstruction, Opcode as DecodedOpcode, Operand,
 };
+use zerocopy::IntoBytes as _;
 
-use crate::{Error, Result, TextSectionInfo, checked_add_u64};
+use crate::{
+    Error, InternalError, Result, TextSectionInfo, checked_add_u64, fixup_phdr_alignment,
+    text_sections,
+};
 
 #[derive(Clone, Copy)]
 pub(crate) struct ScanSections<'a> {
@@ -148,6 +154,211 @@ impl CodeScanRanges {
 pub struct ElfCodeMetadata {
     pub(crate) executable: Vec<TextSectionInfo>,
     pub(crate) identified: Vec<TextSectionInfo>,
+}
+
+impl ElfCodeMetadata {
+    /// Parse an ELF image in aligned storage, relocating an unaligned program
+    /// header table within that storage if required by the object parser.
+    pub fn parse_aligned_in_place(aligned: &mut [u64], byte_len: usize) -> Result<Self> {
+        let bytes = aligned.as_mut_bytes();
+        if byte_len > bytes.len() {
+            return Err(Error::ParseError(
+                "aligned ELF byte length exceeds storage".into(),
+            ));
+        }
+        fixup_phdr_alignment(&mut bytes[..byte_len]);
+        let file = object::File::parse(&bytes[..byte_len])
+            .map_err(|error| Error::ParseError(error.to_string()))?;
+        if file.architecture() != object::Architecture::Aarch64 {
+            return Err(Error::UnsupportedExecutable(
+                "code metadata requires an AArch64 ELF".into(),
+            ));
+        }
+        match elf_code_metadata(&file) {
+            Ok(metadata) => Ok(metadata),
+            Err(InternalError::Public(error)) => Err(error),
+            Err(InternalError::NoTextSectionFound) => Err(Error::UnsupportedExecutable(
+                "AArch64 ELF has no executable code ranges".into(),
+            )),
+            Err(error) => unreachable!("unexpected metadata error: {error:?}"),
+        }
+    }
+
+    /// Project ELF file ranges into one file-backed mapping.
+    pub fn ranges_for_mapping(
+        &self,
+        file_offset: u64,
+        mapping_len: usize,
+    ) -> Result<CodeScanRanges> {
+        Ok(CodeScanRanges {
+            executable: project_file_ranges(&self.executable, file_offset, mapping_len)?,
+            identified: project_file_ranges(&self.identified, file_offset, mapping_len)?,
+        })
+    }
+}
+
+pub(crate) fn elf_code_metadata(
+    file: &object::File<'_>,
+) -> core::result::Result<ElfCodeMetadata, InternalError> {
+    let executable = text_sections(file)?;
+    let identified = code_sections(file, &executable);
+    Ok(ElfCodeMetadata {
+        executable,
+        identified,
+    })
+}
+
+#[expect(
+    clippy::cast_possible_truncation,
+    reason = "the crate requires 64-bit pointers"
+)]
+fn project_file_ranges(
+    ranges: &[TextSectionInfo],
+    mapping_file_offset: u64,
+    mapping_len: usize,
+) -> Result<Vec<Range<usize>>> {
+    let mapping_end = mapping_file_offset
+        .checked_add(mapping_len as u64)
+        .ok_or_else(|| Error::AddressOverflow("mapping file range".into()))?;
+    let mut projected = Vec::new();
+    for range in ranges {
+        let range_end = range
+            .file_offset
+            .checked_add(range.size)
+            .ok_or_else(|| Error::AddressOverflow("ELF code file range".into()))?;
+        let start = range.file_offset.max(mapping_file_offset);
+        let end = range_end.min(mapping_end);
+        if start < end {
+            projected
+                .push((start - mapping_file_offset) as usize..(end - mapping_file_offset) as usize);
+        }
+    }
+    Ok(projected)
+}
+
+/// Builds ranges where metadata positively identifies AArch64 code. Executable
+/// bytes outside these ranges remain eligible for exact mandatory-site scans,
+/// but not heuristic x18 classification.
+fn code_sections(file: &object::File<'_>, executable: &[TextSectionInfo]) -> Vec<TextSectionInfo> {
+    let mut ranges = Vec::new();
+    for symbol in file.symbols().chain(file.dynamic_symbols()) {
+        if symbol.kind() == object::SymbolKind::Text
+            && symbol.size() != 0
+            && let Some(end) = symbol.address().checked_add(symbol.size())
+        {
+            ranges.push((symbol.address(), end));
+        }
+    }
+
+    if let Some(section) = file.section_by_name(".eh_frame")
+        && let Ok(data) = section.uncompressed_data()
+        && let Some(eh_frame_ranges) = eh_frame_ranges(
+            &data,
+            section.address(),
+            file.section_by_name(".text").map(|text| text.address()),
+            file.section_by_name(".got").map(|got| got.address()),
+        )
+    {
+        ranges.extend(eh_frame_ranges);
+    }
+
+    let has_function_metadata = !ranges_to_sections(ranges.clone(), executable).is_empty();
+    if !has_function_metadata {
+        return executable.to_vec();
+    }
+
+    for section in file.sections() {
+        let Ok(name) = section.name() else {
+            continue;
+        };
+        if (matches!(name, ".init" | ".fini" | ".plt" | ".iplt") || name.starts_with(".plt."))
+            && let Some(end) = section.address().checked_add(section.size())
+        {
+            ranges.push((section.address(), end));
+        }
+    }
+
+    ranges_to_sections(ranges, executable)
+}
+
+fn eh_frame_ranges(
+    data: &[u8],
+    eh_frame_address: u64,
+    text_address: Option<u64>,
+    got_address: Option<u64>,
+) -> Option<Vec<(u64, u64)>> {
+    let mut section = gimli::EhFrame::new(data, gimli::LittleEndian);
+    section.set_address_size(8);
+    let mut bases = gimli::BaseAddresses::default().set_eh_frame(eh_frame_address);
+    if let Some(text_address) = text_address {
+        bases = bases.set_text(text_address);
+    }
+    if let Some(got_address) = got_address {
+        bases = bases.set_got(got_address);
+    }
+    let mut entries = gimli::UnwindSection::entries(&section, &bases);
+    let mut ranges = Vec::new();
+    loop {
+        let entry = match entries.next() {
+            Ok(Some(entry)) => entry,
+            Ok(None) => return Some(ranges),
+            Err(_) => return None,
+        };
+        if let gimli::CieOrFde::Fde(partial) = entry {
+            let Ok(fde) = partial.parse(gimli::UnwindSection::cie_from_offset) else {
+                return None;
+            };
+            if fde.len() != 0
+                && let Some(end) = fde.initial_address().checked_add(fde.len())
+            {
+                ranges.push((fde.initial_address(), end));
+            }
+        }
+    }
+}
+
+fn ranges_to_sections(
+    ranges: Vec<(u64, u64)>,
+    executable: &[TextSectionInfo],
+) -> Vec<TextSectionInfo> {
+    let mut clipped = Vec::new();
+    for (start, end) in ranges {
+        for section in executable {
+            let section_end = section.vaddr.saturating_add(section.size);
+            let start = start
+                .max(section.vaddr)
+                .checked_add(3)
+                .map_or(u64::MAX, |start| start & !3);
+            let end = end.min(section_end) & !3;
+            if start < end {
+                clipped.push((start, end, section));
+            }
+        }
+    }
+    clipped.sort_unstable_by_key(|(start, _, _)| *start);
+
+    let mut code: Vec<TextSectionInfo> = Vec::new();
+    for (start, end, section) in clipped {
+        if let Some(previous) = code.last_mut()
+            && start >= previous.vaddr
+            && previous.vaddr.saturating_add(previous.size) >= start
+            && previous.file_offset.saturating_add(start - previous.vaddr)
+                == section.file_offset.saturating_add(start - section.vaddr)
+        {
+            previous.size = end
+                .max(previous.vaddr.saturating_add(previous.size))
+                .saturating_sub(previous.vaddr);
+            continue;
+        }
+        code.push(TextSectionInfo {
+            vaddr: start,
+            file_offset: section
+                .file_offset
+                .saturating_add(start.saturating_sub(section.vaddr)),
+            size: end - start,
+        });
+    }
+    code
 }
 
 // ============================================================
@@ -5026,6 +5237,25 @@ mod tests {
         assert_eq!(sites.len(), 1);
         assert_eq!(sites[0].vaddr, 0x1000);
         assert_eq!(sites[0].kind, PatchKind::Svc);
+    }
+
+    #[test]
+    fn metadata_ranges_are_clipped_aligned_and_merged() {
+        let executable = [TextSectionInfo {
+            vaddr: 0x1000,
+            file_offset: 0x200,
+            size: 0x20,
+        }];
+
+        let ranges = ranges_to_sections(
+            vec![(0x0fff, 0x1006), (0x1004, 0x100d), (0x101e, 0x1024)],
+            &executable,
+        );
+
+        assert_eq!(ranges.len(), 1);
+        assert_eq!(ranges[0].vaddr, 0x1000);
+        assert_eq!(ranges[0].file_offset, 0x200);
+        assert_eq!(ranges[0].size, 0xc);
     }
 
     #[test]
